Fix conditionals that check if image is default before deletion

+++ server/model/userSchema.js
@@ -102,8 +102,8 @@ UserSchema.post('findOneAndDelete', async function() {
 
 async function deleteProfilePic(user) {
     const profilePic = user.profilePic;
-    if (profilePic.bucket !== defaultProfilePic.bucket
-        || profilePic.key !== defaultProfilePic.key) {
+    if (!(profilePic.bucket === defaultProfilePic.bucket
+        && profilePic.key === defaultProfilePic.key)) {
         LOGGER.debug('Deleting profile picture in S3 (bucket: {}, key: {}) for User (_id: {})',
             profilePic.bucket, profilePic.key, user._id);
         await deleteObject({
